Lots of refactoring; Circle, Line and Curve files created; and more
The classes are being abandoned in favor of files containing function sets.

This has been made because the concept of classes in JS is pretty much syntactic sugar and might be even harmful in terms of performance. Besides, every single class method we created thus far was static, so why not functions under a file instead?