Perspective tool finished; yet another big round of refactoring
The whole architecture of the three-dimensional, cube-related projection and perspective tools were refactored so as to maximize the amount of code reuse.

This commit also introduces three perspective techniques per-se (with one, two and three focal points, respectively). The visualization of them is somewhat clunky, and I need to create a means of changing the pixel density of the canvas to improve the visualization.